split invite from categorizables
